fixes failed test cases after codebase restructure (WIP)
