+++ src/canmatrix/ArbitrationId.py
@@ -25,8 +25,9 @@
 # ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE
 # POSSIBILITY OF SUCH DAMAGE.
 
-import typing
 import attr
+import warnings
+import typing
 
 @attr.s
 class ArbitrationId(object):
